pyproject.toml
[tool.poetry]
name = "api-starter"
version = "0.1.0"
description = ""
authors = ["Rash-in <jerry@bytesoffury.com>"]
readme = "README.md"
packages = [{include = "api_starter"}]

[tool.poetry.dependencies]
python = "^3.10"
pipdeptree = "^2.5.0"
colander = "^2.0"
fastapi = "^0.92.0"
gunicorn = "^20.1.0"
uvicorn = "^0.20.0"
loguru = "^0.6.0"
mysqlclient = "^2.1.1"
pytest = "^7.2.1"
setproctitle = "^1.3.2"
python-dotenv = "^1.0.0"
SQLAlchemy = "^2.0.4"
requests = "^2.28.2"
poetry = "^1.3.2"
wheel = "^0.38.4"
setuptools = "^67.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
